# herald-server/src/main.rs
use std::sync::Arc;
use std::time::Duration;

use tracing::{error, info, warn};

use herald_server::config::HeraldConfig;
use herald_server::integrations;
use herald_server::state::{AppState, AppStateBuilder};
use herald_server::store;

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    let args: Vec<String> = std::env::args().collect();

    let config_path = args
        .iter()
        .find(|a| !a.starts_with("--") && *a != &args[0])
        .cloned()
        .unwrap_or_else(|| "herald.toml".to_string());

    let mut config = HeraldConfig::load_or_env(&config_path)?;

    let filter = tracing_subscriber::EnvFilter::try_from_default_env()
        .unwrap_or_else(|_| tracing_subscriber::EnvFilter::new(&config.server.log_level));

    #[cfg(feature = "otel")]
    {
        use opentelemetry::trace::TracerProvider;
        use tracing_subscriber::layer::SubscriberExt;
        use tracing_subscriber::util::SubscriberInitExt;

        let otlp_endpoint = std::env::var("OTEL_EXPORTER_OTLP_ENDPOINT").unwrap_or_default();
        if !otlp_endpoint.is_empty() {
            let exporter = opentelemetry_otlp::SpanExporter::builder()
                .with_tonic()
                .build()
                .expect("failed to create OTLP exporter");
            let tracer_provider = opentelemetry_sdk::trace::SdkTracerProvider::builder()
                .with_batch_exporter(exporter)
                .with_resource(
                    opentelemetry_sdk::Resource::builder()
                        .with_service_name("herald")
                        .build(),
                )
                .build();
            let tracer = tracer_provider.tracer("herald");
            let otel_layer = tracing_opentelemetry::layer().with_tracer(tracer);

            tracing_subscriber::registry()
                .with(filter)
                .with(tracing_subscriber::fmt::layer().json().with_target(true))
                .with(otel_layer)
                .init();

            info!("OpenTelemetry OTLP export enabled → {otlp_endpoint}");
        } else {
            tracing_subscriber::fmt()
                .json()
                .with_env_filter(filter)
                .with_target(true)
                .with_thread_ids(false)
                .with_file(false)
                .with_line_number(false)
                .init();
        }
    }
    #[cfg(not(feature = "otel"))]
    {
        tracing_subscriber::fmt()
            .json()
            .with_env_filter(filter)
            .with_target(true)
            .with_thread_ids(false)
            .with_file(false)
            .with_line_number(false)
            .init();
    }

    // Pull secrets from Keep if HERALD_KEEP_ADDR is set
    config.load_secrets_from_keep().await?;

    // Validate configuration before proceeding
    config.validate()?;

    // Open ShroudB storage — embedded (local WAL) or remote (shared server)
    let db: Arc<herald_server::store_backend::StoreBackend> = match config.store.mode.as_str() {
        "remote" => {
            let addr = config.store.addr.as_deref().ok_or_else(|| {
                anyhow::anyhow!("store.addr is required when store.mode is \"remote\"")
            })?;
            info!(addr = %addr, "connecting to remote ShroudB");
            let remote = shroudb_client::RemoteStore::connect(addr)
                .await
                .map_err(|e| anyhow::anyhow!("remote store connect failed: {e}"))?;
            info!(addr = %addr, "remote store ready");
            Arc::new(herald_server::store_backend::StoreBackend::Remote(remote))
        }
        _ => {
            let data_dir = &config.store.path;
            if let Some(parent) = data_dir.parent() {
                std::fs::create_dir_all(parent)?;
            }
            let storage_config = shroudb_storage::StorageEngineConfig {
                data_dir: data_dir.to_path_buf(),
                ..Default::default()
            };
            let master_key = shroudb_storage::ChainedMasterKeySource::default_chain();
            let storage = Arc::new(
                shroudb_storage::StorageEngine::open(storage_config, &master_key)
                    .await
                    .map_err(|e| anyhow::anyhow!("storage engine failed: {e}"))?,
            );
            let embedded = shroudb_storage::EmbeddedStore::new(storage.clone(), "herald");
            info!(data_dir = %data_dir.display(), "embedded store ready");
            Arc::new(herald_server::store_backend::StoreBackend::Embedded(
                embedded,
            ))
        }
    };

    // Initialize namespaces
    store::init_namespaces(&*db)
        .await
        .map_err(|e| anyhow::anyhow!("namespace init: {e}"))?;

    let bind = config.server.bind.clone();
    let shutdown_timeout = config.server.shutdown_timeout_secs;
    let tls_config = config.tls.clone();

    // Initialize ShroudB integrations — in-process engines for both store backends.
    let mut sentry: Option<Arc<dyn integrations::SentryOps>> = match *db {
        herald_server::store_backend::StoreBackend::Embedded(ref embedded) => {
            let engine = embedded.engine().clone();
            let sentry_store = Arc::new(shroudb_storage::EmbeddedStore::new(engine, "sentry"));
            match integrations::in_process_sentry::InProcessSentryOps::new(sentry_store).await {
                Ok(s) => {
                    info!("Sentry engine initialized (in-process, embedded store)");
                    Some(Arc::new(s) as Arc<dyn integrations::SentryOps>)
                }
                Err(e) => {
                    warn!("in-process Sentry init failed: {e}");
                    None
                }
            }
        }
        herald_server::store_backend::StoreBackend::Remote(_) => {
            let addr = config.store.addr.as_deref().unwrap();
            match shroudb_client::RemoteStore::connect(addr).await {
                Ok(remote_sentry) => {
                    let sentry_store = Arc::new(remote_sentry);
                    match integrations::in_process_sentry::InProcessSentryOps::new(sentry_store)
                        .await
                    {
                        Ok(s) => {
                            info!("Sentry engine initialized (in-process, remote store)");
                            Some(Arc::new(s) as Arc<dyn integrations::SentryOps>)
                        }
                        Err(e) => {
                            warn!("in-process Sentry init failed (remote store): {e}");
                            None
                        }
                    }
                }
                Err(e) => {
                    warn!("failed to connect remote store for Sentry: {e}");
                    None
                }
            }
        }
    };
    // Initialize in-process Courier engine.
    let instance_id = config
        .cluster
        .instance_id
        .clone()
        .unwrap_or_else(|| uuid::Uuid::new_v4().to_string());

    let mut courier: Option<Arc<dyn integrations::CourierOps>> = match *db {
        herald_server::store_backend::StoreBackend::Embedded(ref embedded) => {
            let engine = embedded.engine().clone();
            let courier_store = Arc::new(shroudb_storage::EmbeddedStore::new(engine, "courier"));
            match integrations::in_process_courier::InProcessCourierOps::new(
                courier_store,
                instance_id.clone(),
            )
            .await
            {
                Ok(c) => {
                    info!("Courier engine initialized (in-process, embedded store)");
                    Some(Arc::new(c) as Arc<dyn integrations::CourierOps>)
                }
                Err(e) => {
                    warn!("in-process Courier init failed: {e}");
                    None
                }
            }
        }
        herald_server::store_backend::StoreBackend::Remote(_) => {
            let addr = config.store.addr.as_deref().unwrap();
            match shroudb_client::RemoteStore::connect(addr).await {
                Ok(remote_courier) => {
                    match integrations::in_process_courier::InProcessCourierOps::new(
                        Arc::new(remote_courier),
                        instance_id.clone(),
                    )
                    .await
                    {
                        Ok(c) => {
                            info!("Courier engine initialized (in-process, remote store)");
                            Some(Arc::new(c) as Arc<dyn integrations::CourierOps>)
                        }
                        Err(e) => {
                            warn!("in-process Courier init failed (remote store): {e}");
                            None
                        }
                    }
                }
                Err(e) => {
                    warn!("failed to connect remote store for Courier: {e}");
                    None
                }
            }
        }
    };

    // Initialize in-process Chronicle engine.
    let mut chronicle: Option<Arc<dyn integrations::ChronicleOps>> = match *db {
        herald_server::store_backend::StoreBackend::Embedded(ref embedded) => {
            let engine = embedded.engine().clone();
            let chronicle_store =
                Arc::new(shroudb_storage::EmbeddedStore::new(engine, "chronicle"));
            match integrations::in_process_chronicle::InProcessChronicleOps::new(chronicle_store)
                .await
            {
                Ok(c) => {
                    info!("Chronicle engine initialized (in-process, embedded store)");
                    Some(Arc::new(c) as Arc<dyn integrations::ChronicleOps>)
                }
                Err(e) => {
                    warn!("in-process Chronicle init failed: {e}");
                    None
                }
            }
        }
        herald_server::store_backend::StoreBackend::Remote(_) => {
            let addr = config.store.addr.as_deref().unwrap();
            match shroudb_client::RemoteStore::connect(addr).await {
                Ok(remote_chronicle) => {
                    match integrations::in_process_chronicle::InProcessChronicleOps::new(Arc::new(
                        remote_chronicle,
                    ))
                    .await
                    {
                        Ok(c) => {
                            info!("Chronicle engine initialized (in-process, remote store)");
                            Some(Arc::new(c) as Arc<dyn integrations::ChronicleOps>)
                        }
                        Err(e) => {
                            warn!("in-process Chronicle init failed (remote store): {e}");
                            None
                        }
                    }
                }
                Err(e) => {
                    warn!("failed to connect remote store for Chronicle: {e}");
                    None
                }
            }
        }
    };

    // Remote overrides — if [shroudb] config has addresses, use remote TCP instead of in-process.
    if let Some(ref shroudb) = config.shroudb {
        use integrations::resilient::*;

        if let Some(ref addr) = shroudb.sentry_addr {
            match integrations::sentry::RemoteSentryOps::connect(
                addr,
                shroudb.sentry_token.as_deref(),
            )
            .await
            {
                Ok(s) => {
                    info!(addr = %addr, "connected to Sentry (fail-open circuit breaker)");
                    sentry = Some(Arc::new(ResilientSentry::new(Arc::new(s))));
                }
                Err(e) => warn!(addr = %addr, error = %e, "failed to connect to Sentry"),
            }
        }
        if let Some(ref addr) = shroudb.courier_addr {
            match integrations::courier::RemoteCourierOps::connect(
                addr,
                shroudb.courier_token.as_deref(),
                "default",
            )
            .await
            {
                Ok(c) => {
                    info!(addr = %addr, "connected to Courier (with circuit breaker)");
                    courier = Some(Arc::new(ResilientCourier::new(Arc::new(c))));
                }
                Err(e) => warn!(addr = %addr, error = %e, "failed to connect to Courier"),
            }
        }
        if let Some(ref addr) = shroudb.chronicle_addr {
            match integrations::chronicle::RemoteChronicleOps::connect(
                addr,
                shroudb.chronicle_token.as_deref(),
            )
            .await
            {
                Ok(c) => {
                    info!(addr = %addr, "connected to Chronicle (with circuit breaker)");
                    chronicle = Some(Arc::new(ResilientChronicle::new(Arc::new(c))));
                }
                Err(e) => warn!(addr = %addr, error = %e, "failed to connect to Chronicle"),
            }
        }
    }

    // Register engines based on enabled features
    let engines = herald_server::engines::default_engines();

    let state = AppState::build(AppStateBuilder {
        config,
        db,
        sentry,
        courier,
        chronicle,
        instance_id,
        engines,
    });

    // Always hydrate all tenants from Store
    state.hydrate_tenant_cache().await?;

    // Ensure a default tenant exists (auto-creates on first startup)
    state.bootstrap_default_tenant().await?;
    info!(tenants = state.tenant_cache.len(), "tenants loaded");

    // Hydrate presence overrides from WAL (when presence engine is active)
    #[cfg(feature = "presence")]
    {
        match store::presence_overrides::load_all(&*state.db).await {
            Ok(overrides) => {
                let count = overrides.len();
                state.presence.load_overrides(overrides);
                if count > 0 {
                    info!(count, "loaded presence overrides from WAL");
                }
            }
            Err(e) => {
                warn!("failed to load presence overrides: {e}");
            }
        }
    }

    // Hydrate streams + members into memory
    let all_streams = store::streams::list_all(&*state.db).await?;
    for (tid, stream) in &all_streams {
        let latest_seq = store::events::latest_seq(&*state.db, tid, stream.id.as_str())
            .await
            .unwrap_or(0);
        state
            .streams
            .create_stream(tid, stream.id.as_str(), latest_seq, stream.public);
        if stream.archived {
            state.streams.set_archived(tid, stream.id.as_str(), true);
        }

        // In clustered mode, advance the store-backed sequence counter
        // to match existing events (one-time migration per stream).
        if state.config.cluster.enabled && latest_seq > 0 {
            if let Err(e) =
                store::sequences::advance_to(&*state.db, tid, stream.id.as_str(), latest_seq).await
            {
                warn!(
                    tenant = tid,
                    stream = stream.id.as_str(),
                    "failed to advance sequence counter: {e}"
                );
            }
        }

        if let Ok(members) =
            store::members::list_by_stream(&*state.db, tid, stream.id.as_str()).await
        {
            for member in members {
                state
                    .streams
                    .add_member(tid, stream.id.as_str(), &member.user_id);
            }
        }
    }

    info!(streams = all_streams.len(), "hydrated stream state");

    // Create __presence stream for each tenant (presence engine global broadcast).
    // This is a special stream that receives all presence changes tenant-wide.
    // Clients subscribe to it for app-wide "who's online" without per-user watchlists.
    #[cfg(feature = "presence")]
    {
        use herald_server::ws::connection::PRESENCE_STREAM;
        for tc in state.tenant_cache.iter() {
            let tid = tc.key().clone();
            if !state.streams.stream_exists(&tid, PRESENCE_STREAM) {
                state.streams.create_stream(&tid, PRESENCE_STREAM, 0, true);
                // Persist the stream to WAL
                let stream = herald_core::stream::Stream {
                    id: herald_core::stream::StreamId(PRESENCE_STREAM.to_string()),
                    name: "Presence".to_string(),
                    meta: None,
                    created_at: herald_server::ws::connection::now_millis(),
                    public: true,
                    archived: false,
                };
                let _ = store::streams::insert(&*state.db, &tid, &stream).await;
            }
        }
    }

    // Stats snapshot collector — every 60 seconds
    let stats_state = state.clone();
    let stats_handle = tokio::spawn(async move {
        loop {
            tokio::time::sleep(Duration::from_secs(60)).await;
            let connections = stats_state.connections.total_connections() as u64;
            let events = stats_state
                .metrics
                .events_published
                .load(std::sync::atomic::Ordering::Relaxed);
            let streams = stats_state.streams.stream_count() as u64;
            let auth_failures = stats_state
                .metrics
                .ws_auth_failures
                .load(std::sync::atomic::Ordering::Relaxed);
            stats_state
                .event_bus
                .record_snapshot(connections, events, streams, auth_failures);

            // Per-tenant snapshots
            for entry in stats_state.tenant_cache.iter() {
                let tid = entry.key();
                let t_connections = stats_state.connections.tenant_connection_count(tid) as u64;
                let t_events = stats_state.tenant_events_published(tid);
                let t_webhooks = stats_state.tenant_webhooks_sent(tid);
                let t_streams = stats_state.streams.tenant_stream_count(tid) as u64;
                stats_state.event_bus.record_tenant_snapshot(
                    tid,
                    t_connections,
                    t_events,
                    t_webhooks,
                    t_streams,
                );
            }
        }
    });

    // Backplane: cross-instance fanout via ShroudB subscriptions
    let backplane_handle = if state.config.cluster.enabled {
        info!(
            instance_id = %state.instance_id,
            "cluster mode enabled — starting backplane consumer"
        );
        Some(herald_server::backplane::spawn(state.clone()))
    } else {
        None
    };

    // Engine timer tasks (typing TTL expiry, presence sweep, etc.)
    let engine_timer_handles: Vec<_> = state
        .engines
        .timers()
        .into_iter()
        .map(|task| {
            let timer_state = state.clone();
            info!(task = task.label, "starting engine timer");
            tokio::spawn(async move {
                loop {
                    tokio::time::sleep(task.interval).await;
                    (task.run)(timer_state.clone()).await;
                }
            })
        })
        .collect();

    // TTL cleanup
    let cleanup_db = state.db.clone();
    let cleanup_handle = tokio::spawn(async move {
        loop {
            tokio::time::sleep(Duration::from_secs(3600)).await;
            let now = herald_server::ws::connection::now_millis();
            match store::events::delete_expired(&*cleanup_db, now).await {
                Ok(n) if n > 0 => info!(deleted = n, "TTL cleanup: pruned expired events"),
                Err(e) => warn!("TTL cleanup error: {e}"),
                _ => {}
            }
        }
    });

    // Shutdown signal
    let (shutdown_tx, shutdown_rx) = tokio::sync::watch::channel(false);

    let app = herald_server::http::router(state.clone());

    info!(bind = %bind, "starting Herald");

    let make_shutdown = |rx: tokio::sync::watch::Receiver<bool>| async move {
        let mut rx = rx;
        while !*rx.borrow() {
            if rx.changed().await.is_err() {
                break;
            }
        }
    };

    if let Some(ref tls) = tls_config {
        let rustls_config =
            axum_server::tls_rustls::RustlsConfig::from_pem_file(&tls.cert_path, &tls.key_path)
                .await?;

        let addr: std::net::SocketAddr = bind.parse()?;
        let server_handle = tokio::spawn(async move {
            if let Err(e) = axum_server::bind_rustls(addr, rustls_config)
                .serve(app.into_make_service())
                .await
            {
                error!("server error: {e}");
            }
        });

        tokio::signal::ctrl_c().await?;
        info!("shutdown signal received, draining...");
        let _ = shutdown_tx.send(true);

        state
            .connections
            .broadcast_all(&herald_core::protocol::ServerMessage::error(
                None,
                herald_core::error::ErrorCode::Internal,
                "server shutting down",
            ));

        tokio::time::sleep(Duration::from_millis(500)).await;
        cleanup_handle.abort();
        stats_handle.abort();
        if let Some(ref h) = backplane_handle {
            h.abort();
        }
        for h in &engine_timer_handles {
            h.abort();
        }

        let _ = tokio::time::timeout(Duration::from_secs(shutdown_timeout), async {
            let _ = server_handle.await;
        })
        .await;
    } else {
        let listener = tokio::net::TcpListener::bind(&bind).await?;

        let shutdown = shutdown_rx.clone();
        let server_handle = tokio::spawn(async move {
            if let Err(e) = axum::serve(listener, app)
                .with_graceful_shutdown(make_shutdown(shutdown))
                .await
            {
                error!("server error: {e}");
            }
        });

        tokio::signal::ctrl_c().await?;
        info!("shutdown signal received, draining...");
        let _ = shutdown_tx.send(true);

        state
            .connections
            .broadcast_all(&herald_core::protocol::ServerMessage::error(
                None,
                herald_core::error::ErrorCode::Internal,
                "server shutting down",
            ));

        tokio::time::sleep(Duration::from_millis(500)).await;
        cleanup_handle.abort();
        stats_handle.abort();
        if let Some(ref h) = backplane_handle {
            h.abort();
        }
        for h in &engine_timer_handles {
            h.abort();
        }

        let _ = tokio::time::timeout(Duration::from_secs(shutdown_timeout), async {
            let _ = server_handle.await;
        })
        .await;
    }

    info!("shutdown complete");
    Ok(())
}