# openapi.yaml
openapi: 3.1.0
info:
  title: Herald API
  description: Persistent realtime event streams with built-in authorization.
  license:
    name: Proprietary
  version: 2.0.0
servers:
- url: /
  description: Local server
paths:
...
  /streams:
    get:
      tags:
      - streams
      operationId: list_streams
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: List of streams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    post:
      tags:
      - streams
      operationId: create_stream
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStreamRequest'
        required: true
      responses:
        '201':
          description: Stream created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stream'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Stream limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Stream already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
  /streams/{id}/members:
    get:
      tags:
      - members
      operationId: list_members
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
    post:
      tags:
      - members
      operationId: add_member
      parameters:
      - name: id
        in: path
        description: Stream ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
        required: true
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - basic_auth: []
      - bearer_auth: []
...
components:
  schemas:
    AddMemberRequest:
      type: object
      required:
      - user_id
      properties:
        role:
          type:
          - string
          - 'null'
        user_id:
          type: string
...
    CreateStreamRequest:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
        meta: {}
        name:
          type: string
        public:
          type:
          - boolean
          - 'null'
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    Member:
      type: object
      required:
      - stream_id
      - user_id
      - joined_at
      properties:
        joined_at:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/Role'
        stream_id:
          type: string
        user_id:
          type: string
    MemberListResponse:
      type: object
      required:
      - members
      - total
      - limit
      - offset
      properties:
        limit:
          type: integer
          minimum: 0
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
...
    Role:
      type: string
      enum:
      - owner
      - admin
      - member
...
    Stream:
      type: object
      required:
      - id
      - name
      - created_at
      properties:
        archived:
          type: boolean
        created_at:
          type: integer
          format: int64
        id:
          $ref: '#/components/schemas/StreamId'
        meta: {}
        name:
          type: string
        public:
          type: boolean
    StreamId:
      type: string
    StreamListResponse:
      type: object
      required:
      - streams
      - total
      - limit
      - offset
      properties:
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        streams:
          type: array
          items:
            $ref: '#/components/schemas/Stream'
        total:
          type: integer
          minimum: 0
...
  securitySchemes:
    admin_auth:
      type: http
      scheme: bearer
    basic_auth:
      type: http
      scheme: basic
    bearer_auth:
      type: http
      scheme: bearer
security:
- basic_auth: []
- bearer_auth: []
- admin_auth: []
...